move name class to the common package
