+++ src/seedu/addressbook/common/Name.java
@@ -1,0 +1,64 @@
+package seedu.addressbook.common;
+
+import seedu.addressbook.data.exception.IllegalValueException;
+
+import java.util.Arrays;
+import java.util.List;
+
+/**
+ * Represents a Person's name in the address book.
+ * Guarantees: immutable; is valid as declared in {@link #isValidName(String)}
+ */
+public class Name {
+
+    public static final String EXAMPLE = "John Doe";
+    public static final String MESSAGE_NAME_CONSTRAINTS = "Person names should be spaces or alphanumeric characters";
+    public static final String NAME_VALIDATION_REGEX = "[\\p{Alnum} ]+";
+
+    public final String fullName;
+
+    /**
+     * Validates given name.
+     *
+     * @throws IllegalValueException if given name string is invalid.
+     */
+    public Name(String name) throws IllegalValueException {
+        name = name.trim();
+        if (!isValidName(name)) {
+            throw new IllegalValueException(MESSAGE_NAME_CONSTRAINTS);
+        }
+        this.fullName = name;
+    }
+
+    /**
+     * Returns true if a given string is a valid person name.
+     */
+    public static boolean isValidName(String test) {
+        return test.matches(NAME_VALIDATION_REGEX);
+    }
+
+    /**
+     * Retrieves a listing of every word in the name, in order.
+     */
+    public List<String> getWordsInName() {
+        return Arrays.asList(fullName.split("\\s+"));
+    }
+
+    @Override
+    public String toString() {
+        return fullName;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Name // instanceof handles nulls
+                && this.fullName.equals(((Name) other).fullName)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return fullName.hashCode();
+    }
+
+}
